refold-ease: do not change ease factors directly through db
By modifying the ease factors of cards directly through mw.col.db, Anki
doesn't know that card properties have been changed, which in turn
requires a force-sync. Force-syncs are inconvenient and it's simply much
more efficient to use the recommended workflow for modifying cards.

This is effectively an identical patch to the one applied to
ResetEase[1].

[1]: https://github.com/AnKingMed/Reset-Ease/pull/4

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>